Add service test functionality and refactor code
